+++ CHANGELOG.md
@@ -2,12 +2,20 @@
 
 All notable changes to the "quarkus-vsc" extension will be documented in this file.
 
-## 0.3.0 – 2019-07-10
+## 0.4.0 – 2019-06-12
+
+-   Added config option for metrics
+-   Added config option for tracing
+-   Added config option for healthcheck
+-   Added config option for kubernetes support
+-   Fixed issues with config generation (current dir vs new dir)
+
+## 0.3.0 – 2019-06-11
 
 -   Minor text fixes
 -   Updated Readme
 
-## 0.2.0 – 2019-07-10
+## 0.2.0 – 2019-06-10
 
 -   Fixed issues with adding extensions
 -   Updated extensions list